TuanLA_20231102_Add_Check_Banned_Word

+++ app/src/main/java/com/anhtuan/bookapp/config/Constant.java
@@ -13,7 +13,7 @@ public class Constant {
     // ip tuan
 //    public static final String IP_SERVER = "http://192.168.130.143:8080/";
 
-    public static final int A_DAY = 86400000;
+    public static final long A_DAY = 86400000;
 
     public static class AddBookType{
         public static final int MEMBER_ADD = 1;
